Refactor stream event to its own mod
